Add C++23 support in the configure
Clang main line of March or so (i.e. clang 15) supports the multi arg
bracket operator. However it does not have the feature macro
implemented, so mdspan will not automatically use it.
To use bracket operator configure with this:

-DCMAKE_CXX_FLAGS="-DMDSPAN_USE_BRACKET_OPERATOR=1 -Wno-pre-c++2b-compat"

Note the warning is there, because C++23 was not released yet. You would
get a lot of them if you don't turn this off, when doing -pedantic
builds.