submdspan: slice integral_constant
If submdspan is given an integral_constant as a slice specifier,
pass it along into the offset computation as a compile-time value.

This is perhaps less useful than the tuple of two integral_constant
case, which can turn a dynamic extent into a static extent.  However,
it's not hard to do.