aligned_accessor: Address review feedback
Don't use a single allocation for aligned and unaligned access
any more.  Instead, separate those allocations, and encapsulate
them more.  Add comments to make it clear why we want a
"deliberately unaligned array" for benchmarks and for studying
code generation.  Also, make sure that the elements of the arrays
are filled before we start computing with them.