Fix P3663 benchmark and expand
Make sure that the benchmark gets the MDSPAN_ENABLE_P3663
definition (it wasn't before).

Add new benchmarks that make it harder for the compiler
to optimize away code.  (They actually change all the values
in the mdspan and check that the changes are mathematically
correct.)

In benchmark, for the slice type that's convertible to
full_extent_t, make the conversion operator noexcept.